// src/shared/types/generated/application/NavigationSnapshot.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FolderNavigationItem } from "./FolderNavigationItem";
import type { SmartFolderNavigationItem } from "./SmartFolderNavigationItem";

export type NavigationSnapshot = { folders: Array<FolderNavigationItem>, smart_folders: Array<SmartFolderNavigationItem>, revision: number, };